+++ .gitignore
@@ -190,3 +190,4 @@ git_stats/*
 StyleCop.Cache
 bin
 obj
+.vs/ExcelToWord/FileContentIndex
