+++ .vs/ExcelToWord/v17/DocumentLayout.json
@@ -1,0 +1,41 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "C:\\Users\\\u0415\u0433\u043E\u0440\\source\\E2W\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{F1BEDCF0-FAA1-4024-B078-1D6C8F871C9D}|ExcelToWord\\ExcelToWord.csproj|c:\\users\\\u0435\u0433\u043E\u0440\\source\\e2w\\exceltoword\\findandreplaceclass.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{F1BEDCF0-FAA1-4024-B078-1D6C8F871C9D}|ExcelToWord\\ExcelToWord.csproj|solutionrelative:exceltoword\\findandreplaceclass.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    }
+  ],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": [
+        {
+          "DockedWidth": 200,
+          "SelectedChildIndex": 1,
+          "Children": [
+            {
+              "$type": "Bookmark",
+              "Name": "ST:128:0:{75188d03-9892-4ae2-abf1-207126247ce5}"
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "FindAndReplaceClass.cs",
+              "DocumentMoniker": "C:\\Users\\\u0415\u0433\u043E\u0440\\source\\E2W\\ExcelToWord\\FindAndReplaceClass.cs",
+              "RelativeDocumentMoniker": "ExcelToWord\\FindAndReplaceClass.cs",
+              "ToolTip": "C:\\Users\\\u0415\u0433\u043E\u0440\\source\\E2W\\ExcelToWord\\FindAndReplaceClass.cs",
+              "RelativeToolTip": "ExcelToWord\\FindAndReplaceClass.cs",
+              "ViewState": "AQIAAAAAAAAAAAAAAAAAAAYAAAAAAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-01-11T08:46:54.917Z",
+              "EditorCaption": ""
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
