+++ ExcelToWord/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.Office.Interop.Excel" version="15.0.4795.1000" targetFramework="net472" />
-  <package id="Microsoft.Office.Interop.Word" version="15.0.4797.1003" targetFramework="net472" />
+  <package id="DocumentFormat.OpenXml" version="3.2.0" targetFramework="net472" />
+  <package id="DocumentFormat.OpenXml.Framework" version="3.2.0" targetFramework="net472" />
   <package id="WindowsAPICodePack" version="1.1.1" targetFramework="net472" />
 </packages>
